CI: switch to GitHub Actions (#93)
* CI: switch to GitHub Actions - step 1: sniff stage

This commit:
* Adds a GH Actions workflow for the code style related CI checks.
* Removes that part of the `.travis.yml` configuration.
* Adds a "Build Status" badge in the Readme to use the results from this particular GH Actions runs.

Notes:
Builds will run on all pushes and on pull requests, with the exception of those just modifying files which are irrelevant to this workflow.

* PHPUnit: make WP install script executable

Change the file permissions for the `phpunit/install.sh` to allow execution of the script.

* CI: switch to GitHub Actions - step 2: tests

This commit:
* Adds a GH Actions workflow for the test runs which were previously run on Travis.
* Removes the, now redundant, `.travis.yml` configuration.
* Updates the `.gitattributes` file.
* Updates the "Build Status" badge in the Readme to show the results from the GH `Test` Actions runs.

Notes:
1. In the build matrix on Travis, PHP 7.0 - 7.4 were tested against the "latest" WP version and a few select WP versions were tested against lower PHP versions.
    In the proposed workflow for GH Actions, the matrix is made much more complete with:
    - PHP 5.6 up to 8.0 being tested against both the lowest (`3.7`) and highest (`latest`) officially supported WP version for this plugin.
    - PHP 5.3 up to 5.5 being tested against the lowest (`3.7`) and the highest (`5.1`) WP version which will allow still run on these PHP versions.
    - Selectively testing against some interim WP versions.
2. For the MySQL version, we need to take into account when support for which MySQL version was added to WP.
    - WP 4.5 is the first WP version which properly supports MySQL 5.7.
    - WP 5.4 is the first WP version which largely supports MySQL 8.0.
3. It appears that the PHP 5.6 and 7.0 images do not natively come with the `mysql` or `mysqli` extensions.
    For that reason, the extension is explicitly added.
    Refs:
    * https://github.com/shivammathur/setup-php/wiki/Php-extensions-loaded-on-ubuntu-18.04#php-56
    * https://github.com/shivammathur/setup-php/wiki/Php-extensions-loaded-on-ubuntu-18.04#php-70
4. As these tests are integration tests with the WP native test suite, we need to take the PHPUnit version supported by WP into account for all supported WP versions.
    - Originally only PHPUnit 4.x and 5.x were supported.
    - As of WP 5.1, PHPUnit 6.x and 7.x are also supported.
    - With WP 5.2 dropping support for PHP < 5.6, support for PHPUnit 4.x was also dropped.
    - On PHP >= 8.0, PHPUnit 7.5+ installed via Composer using `--ignore-platform-reqs` is needed, no matter what.

**Known issue**:
It appears that the MySQL version being set in the `services -> image` part is not being respected as can be verified via the `Verify mysql version` step.
No matter what I do, MySQL 5.7 will be installed. This is, at least in part, the reason some of the test runs against WP 3.7 are currently failing.

* GHActions: minor tweaks based on PR review

* GH Actions: allow for manually triggering a workflow

Triggering a workflow for a branch manually is not supported by default in GH Actions, but has to be explicitly allowed.

Ref: https://github.blog/changelog/2020-07-06-github-actions-manual-triggers-with-workflow_dispatch/

* GH Actions: report CS violations in the PR

The cs2pr tool will allow to display the results from an action run in checkstyle format in-line in the PR code view, which should improve usability of the workflow results.

Ref: https://github.com/staabm/annotate-pull-request-from-checkstyle

* GH Actions: fix running against various MySQL versions

These seemingly minor tweaks allow for installing different MySQL versions in the local VM depending on the variables set in the matrix.

This fixes the previously seen test failures when the tests were being run against MySQL 5.7 with an incompatible WP version.

* GH Actions: limit the use of MySQL 8.0 to PHP 7.4 or higher

... as based on the test output, there are issues in WP Core which prevent the tests from being run against MySQL 8.0 using PHP 7.3 or lower.

These issues have been reported upstream. See: https://core.trac.wordpress.org/ticket/52496

* GH Actions: don't test against PHP nightly

WP Core limits the PHPUnit version used for the test suite to PHPUnit 5.7 - 7.5, neither or which are still supported by PHPUnit.

These PHPUnit versions contain an incompatibility with PHP 8.1 which will not be fixed anymore (as the PHPUnit versions are no longer supported).

Until a solution for this has been found and applied to the WP Core test suite - be it widening the supported PHPUnit version range or otherwise -, running the tests against PHP 8.1/nightly is useless as they will fail on the PHPUnit incompatibility without ever running the plugin tests.

Co-authored-by: jrfnl <jrfnl@users.noreply.github.com>
Co-authored-by: Dominik Schilling <dominikschilling+git@gmail.com>